Allows delayed removal for bufer

+++ BuferMAN.ContextMenu/DeleteClipMenuItem.cs
@@ -13,6 +13,8 @@ namespace BuferMAN.ContextMenu
         private readonly IClipboardBuferService _clipboardBuferService;
         private BuferViewModel _bufer;
         private Button _button;
+        private Timer _timer = new Timer();
+        private bool _deleted;
 
         public DeleteClipMenuItem(IClipboardBuferService clipboardBuferService, BuferViewModel bufer, Button button)
         {
@@ -20,15 +22,45 @@ namespace BuferMAN.ContextMenu
             this._bufer = bufer;
             this._button = button;
             this.Text = Resource.DeleteClipMenuItem;
-            this.Shortcut = Shortcut.Del;
-            this.Click += this._DeleteBufer;
+
+            this.MenuItems.Add(new MenuItem(Resource.DeleteBuferNowMenuItem, this._DeleteBufer, Shortcut.Del));
+            this.MenuItems.Add(new MenuItem(Resource.DeleteBuferIn10MinutesMenuItem, this._GetDeleteBuferHandler(10)));
+            this.MenuItems.Add(new MenuItem(Resource.DeleteBuferIn30MinutesMenuItem, this._GetDeleteBuferHandler(30)));
+        }
+
+        private EventHandler _GetDeleteBuferHandler(int deleteInMinutes)
+        {
+            return (object sender, EventArgs e) =>
+            {
+                this._deleted = false;// TODO : bad code. Remove after context menu regeneration on any button creation (recreation)
+                _timer.Interval = deleteInMinutes * 60 * 1000;
+                _timer.Tick += this._RemoveBufer;
+                _timer.Start();
+            };
+        }
+
+        private void _RemoveBufer(object sender, EventArgs e)
+        {
+            if (this._timer.Enabled)
+            {
+                this._timer.Stop();
+            }
+
+            if (!this._deleted)
+            {
+                this._clipboardBuferService.RemoveBufer(this._bufer.ViewId);
+                WindowLevelContext.Current.RerenderBufers();
+
+                this._deleted = true;
+                this._timer.Dispose();
+            }
         }
 
         private void _DeleteBufer(object sender, EventArgs e)
         {
+            this._deleted = false;// TODO : bad code. Remove after context menu regeneration on any button creation (recreation)
             var tabIndex = this._button.TabIndex;
-            this._clipboardBuferService.RemoveBufer(this._bufer.ViewId);
-            WindowLevelContext.Current.RerenderBufers();
+            this._RemoveBufer(sender, e);
 
             tabIndex = this._GetNearestTabIndex(tabIndex);
